requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0
pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
pandas>=1.3.0
datasets>=2.0.0
clip-by-openai>=1.0
lpips>=0.1.4
scikit-image>=0.19.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.0
omegaconf>=2.1.0
hydra-core>=1.1.0
wandb>=0.12.0
tensorboard>=2.8.0
gradio>=3.0.0
streamlit>=1.20.0 